+++ src/translations/en/tag-manager.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Tags for Craft CMS
+ *
+ * @link      https://ethercreative.co.uk
+ * @copyright Copyright (c) 2019 Ether Creative
+ */
+
+return [
+	'Usage' => 'Usage',
+	'Enable Usage' => 'Enable Usage',
+	'Will enable the usage column in the CP (has no effect on regular tag queries, may be slow on larger sites).' =>
+		'Will enable the usage column in the CP (has no effect on regular tag queries, may be slow on larger sites).',
+];